Add PostgreSQL database support and initial migration
Integrated PostgreSQL with the infrastructure, including configurations and migrations for database schema initialization. Updated relevant entities and dependencies to align with the database changes.

+++ src/TelegramBot.ApiService/Domain/Entities/ChatMailingList.cs
@@ -2,8 +2,8 @@
 
 public class ChatMailingList
 {
-    public long ChatId { get; set; }
-    public long MailingListId { get; set; }
+    public int ChatId { get; set; }
+    public int MailingListId { get; set; }
     public DateTime AddedAt { get; set; }
 
     // Навигационные свойства
